test phase 1

+++ NodeManager/src/main/scala/ca/usask/agents/macrm/nodemanager/utils/RealServerState.scala
@@ -19,7 +19,7 @@ object RealServerState extends ServerState {
 
     def getServerResource(): Resource = null
 
-    def createContainer(userId: Int, jobId: Long, taskIndex: Int, size: Resource): Option[Int] = None
+    def createContainer(userId: Int, jobId: Long, taskIndex: Int, size: Resource): Option[Long] = None
 
-    def killContainer(containerId: Long): Option[Long] = None
+    def killContainer(containerId: Long): Option[Int] = None
 }
