+++ NodeManager/src/main/scala/ca/usask/agents/macrm/nodemanager/utils/ServerState.scala
@@ -21,9 +21,9 @@ trait ServerState {
 
     def initializeSimulationServer(resource: Resource, capability: List[Constraint]): Boolean
 
-    def createContainer(userId: Int, jobId: Long, taskIndex: Int, size: Resource): Option[Int]
+    def createContainer(userId: Int, jobId: Long, taskIndex: Int, size: Resource): Option[Long]
 
-    def killContainer(containerId: Long): Option[Long]
+    def killContainer(containerId: Long): Option[Int]
 
 }
